+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/zgreathouse/TournamentLocator#readme",
   "dependencies": {
+    "cookie-session": "^2.0.0-beta.3",
     "express": "^4.16.2",
     "mongoose": "^4.13.9",
     "nodemon": "^1.14.11",
